outsource the navbar element
