+++ src/pages/New.js
@@ -1,12 +1,18 @@
 import styled from "styled-components";
 import { useNavigate} from "react-router-dom"; 
+import NavBarNewOrder from "../components/NavBarNewOrder";
+
+
 export default function New(){
     const navigate = useNavigate(); 
     return(
+        <>
         <NewEntry>
             <button onClick={() => {navigate("/newOrder")}}>neuer Eintrag</button>
             <button onClick={() => {navigate("/newCustomer")}}>neuer Kunde</button>
         </NewEntry>
+        <NavBarNewOrder page={"new"}/>
+        </>
     );
 }
 
@@ -19,5 +25,7 @@ export const NewEntry=styled.div`
     button{
         margin: 10px;
         flex: auto;
+        border-radius:10px;
+        border: 1px solid;
     }
 `
